missing comment added

+++ payments/kirki-redsys/src/RedsysTransactionBuilder.php
@@ -36,7 +36,7 @@ class RedsysTransactionBuilder
             'DS_MERCHANT_TRANSACTIONTYPE' => RedsysConstant::TRANSACTION_TYPE_AUTHORIZATION,
             'DS_MERCHANT_MERCHANTDATA' => $this->order->uuid,
             'DS_MERCHANT_ORDER' => $this->generate_order_number(),
-            'DS_MERCHANT_CURRENCY' => (int) Money::get_currency_numeric_code($this->order->currency_code),
+            'DS_MERCHANT_CURRENCY' => Money::get_currency_numeric_code($this->order->currency_code),
             'DS_MERCHANT_AMOUNT' => (int) $this->order->invoiced_total,
             'DS_MERCHANT_EMV3DS' => wp_json_encode($this->build_emv_3ds()),
         ];
